preprocessor: fix lookahead token stashing for func-like macro id's
When encountering an id token currently defined as a func-like macro,
the preprocessor has to look ahead for an opening parenthesis in order
to decide whether this is actually a macro invocation or not.

However, in case it's not, the next non-empty/-space token doesn't get
properly queued up for later use.

Fix this.

Signed-off-by: Nicolai Stange <nstange@suse.de>