optimize used_macros' und used_macro_undefs' sizes
A pp_token instance contains a used_macros and a used_macro_undefs
instance and each carries a std::set which is huge.

Furthermore, the likely case is that both, used_macros and
used_macro_undefs are empty, so that the two std::set's are unused most of
the time.

Implement class cow_set which is a copy-on-write wrapped std::set.

It allows for COW-style data sharing and in the empty case, it occupies
only the size of a pointer.

Let used_macros and used_macro_undefs use that instead.

Signed-off-by: Nicolai Stange <nstange@suse.de>