preprocessor: treat macro dependencies and reevaluation history separately
Currently, a pp_token's ->used_macros serves two purposes:
1. It keeps track of a token's dependencies on #define's
2. It keeps track of the macro (re-)evaluation history in order to be
   able to detect recursion.

However, conceptually, these are two different things and in particular,
the corresponding sets of macros must differ (which isn't possible as
it stands).

Consider a macro invocation during reevaluation. Strictly speaking,
all macro expansions that have taken part in the creation of the
invocation's accompanyings tokens must be recorded as a dependency to
the corresponding #defines: it is important whether any of them expands
to a argument delimiter, i.e. '(', ',', ')', _or not_.

Currently, we track only the those that do expand to some such delimiter
but not the ones that don't.

Thus, we want to record as a dependency at each token stemming from a
macro expansion each macro expansion which has been involved in the
creation of the tokens making up that invocation.

On the other hand, macro reevalutation must stop once recursion is
detected and the above scheme is too coarse for tracking this.
Consider for example

  #define a b
  #define c(d) a
  #define e(f) c(f)

The correct expansion is e(a) => e(b) => c(b) => a => b since the initial
a => b parameter expansion is not involved at all in the creation of
the intermediate token a stemming from expansion of macro c() => a.

Hence track at each token the dependencies on previous #define's
separately from the expansion history.

The former is kept in a pp_token's _used_macros just as before and for
the latter, a new ->_expansion_history member is introduced.

Since ->_expansion_history is only needed from the preprocessor
for avoiding recursion, it is cleared once a token leaves the
preprocessor.

Similarly, an unexpanded macro argument's ->used_macros() is moved to the
macro invocation's ->_used_macros_base and thus, added to each pp_token
generated from that invocation.

Signed-off-by: Nicolai Stange <nstange@suse.de>