add support for pre-inclusion
gcc has a -include <file> option which allows for <file> being processed early.
We'll need a similar mechanism for pre-including macro definitions which are to
be provided by the compiler.

Let preprocessor not only accept a single header_inclusion_root, but
many thereof. Each will be processed in turn.

Streamline the respective data members and constructors of gnuc_parser_driver
and preprocessor.

Introduce a --include="<file>" command line option to the 'cparse'
test utility. Implement searching for those files from the $PWD in
header_resolver.

Signed-off-by: Nicolai Stange <nstange@suse.de>