parser: recheck for typedef id after for construct
For the following code,

  typedef int T;
  for (int T;;) ;
  T a;

the parser incorrectly classifies the last occurence of T as a
TOK_IDENTIFIER rather as a TOK_TYPEDEF_IDENTIFIER.

The reason is that the Bison parser will have read the lookahead token
already when it reduces the for construct.

Fix this by introducing special case code to the for rule reduction
code which checks for this situation and fixes this up if needed.

Signed-off-by: Nicolai Stange <nstange@suse.de>