ast: resolve: let selection and iteration statements open new scopes
By C99, if, if-else, switch, for, while and do-while statements open
a new (block) scope as do their substatements.

This matters when it comes to struct or union definitions as those can
be defined inside of a sizeof expression, for example.

Make the _id_resolver enter a new scope once it encounters any of the
above. Furthermore, make _id_resolver enter a new scope for any
statement whose parent is any of the above.

Signed-off-by: Nicolai Stange <nstange@suse.de>