parser: fix parameter declarator parsing when redeclaring typedef ids
When there's an ambiguity between interpreting an id in a parameter
declaration as a typedef id or not, then C99 mandates that
the former wins.

Consider

  typedef int x;
  void proc(int (*x)(void)) {}

The x in this function's prototype must not be parsed as a type name
but it currently happens to be.

Fix this.

Be careful to still parse the parameter in

  typedef int x;
  void proc(int (x)) {}

as a typename, i.e. a "function taking an argument of type x and returning
an int".

Signed-off-by: Nicolai Stange <nstange@suse.de>