ast: introduce double dispatch mechanism
The dynamic_cast<>ery found all over the place in ast_resolve.c
is awful, not very C++ish and, depending on the number of casts tried in
and if-else chain, inefficient.

The proper C++ way of handling these kind of situations is to use the
double dispatch idiom,

Introduce a pure virtual base class ast::processor with one virtual
operator() for each leaf ast entity type. These are supposed to get
redefined in derived classes to do actual work for a particular entity
type.

Within the AST entity class hiearchy, introduce one virtual
_process(processor&) function at each leaf type. These dispatch the call
to the respective processor::operator() overloaded for that AST entity
type.

Signed-off-by: Nicolai Stange <nstange@suse.de>