ast: migrate ast::for_each_parent() to double dispatching scheme
When starting a ::for_each_parent(callable_type&&) walk from any ast
entity, that entity type's set closure on ::parent_types is determined.
In order to invoke the callable with an ancestor's dynamic type, a
dynamic_cast<> if-else chain is tried on each member of that set
closure, c.f type_set::cast_and_call().

Since the number of tried types can be large, the if-else chain's length
is basically unbound -- which is bad.

Make use of the double dispatch mechanism recently introduced.

Note however, that each use of this double dispatch mechanism requires
a class derivation from ast::processor. ast::processor has 106 virtual
members by now and thus, any class derivation adds another vtable
slightly less than 1kB in size to the binary.

Introduce a parameter, 'double_dispatch_threshold', currently set to
eight. If the set's cardinality of to be tried types is less than this,
do not use the double dispatch mechanism but fall back to the if-else
chains of dynamic_cast trials, type_set::cast_and_call() that is.

Lax benchmarks show a runtime improvement of ~2.5% for a 'make check'.

Also, rename ->for_each_parent() to ->for_each_ancestor() because
it's really the ancestors which are walked.

Rename the 'mask_type_set' template parameter to 'boundary_type_set'
because it describes its purpose better, namely to stop the walk
immediately after any ast entity of a type therein has been encounterd.

In the light of this, require the callable to return void (rather than
false) for these types and enforce the walk termination from a wrapper.

Signed-off-by: Nicolai Stange <nstange@suse.de>