callables_wrapper: provide no_default_action implementation
Sometimes it is necessary that a callables_wrapper does not overload
operator() with a default_action in case the arguments' types aren't
handled by the underlying callables.

Provide a no_default_action template to be used for callables_wrapper's
default_action template parameter. If given, make callables_wrapper not to
emit any operator() overloads for argument types not recognized.

Signed-off-by: Nicolai Stange <nstange@suse.de>