type_set: make ::cast_and_call() invoke the callable on failure
Currently, type_set<...>::cast_and_call() will try to dynamic_cast<>
to any of the types given and call a dummy marked with
__builtin_unreachable() on failure.

As ::cast_and_call() was used only from the for_each_ancestor() primitive
with the set closure on the declared parent types, this had been
the correct.

However, for_each_dfs_*() will make use of ::cast_and_call() as well
and there, a default action must be invoked on casting failure.

In order to prepare for this, let ::cast_and_call() invoke the callable
with the original (non-casted) type.

To support the old for_each_ancestor() use case, implement a
default_action_unreachable template. In order not to accidentally
provide an operator() for any of the ancestor types (and thus hide
programming errors), let it take a mask type set for which an operator()
will not be defined.

Adopt for_each_ancestor() accordingly.

Signed-off-by: Nicolai Stange <nstange@suse.de>