ast: introduce a proper entity for 'offset_member_designator'
Currently the second parameter of GCC's __builtin_offsetof() extension is
modeled to be an expr. This has been so because the
'offset_member_designator' grammar production (c.f. the GCC manual),
happened to be a subset of the C99 'expr' instruction.

However, representing the member designator by an expr instance is
cumbersome for later stages.

Introduce a proper 'offset_member_designator' AST entity. Make the
parser create it.

As a bonus, make the resolver cease to accept resolving failures for
the special case of expr_id's being childs of a expr_builtin_offsetof.

Signed-off-by: Nicolai Stange <nstange@suse.de>