ast resolver: streamline representation of expr_id::resolved
Currently, the representation of an expr_id to an init/function parameter
declaration or to a function definition takes an indirection:
the expr_id links to a direct_declarator_id which in turn has got
a context assigned to it.

That context then refers to the enclosing init/function parameter
declaration or function definition. However, a direct_declarator_id's
context can also be a struct_declarator entity and an expr_id
clearly can't get resolved to these.

This indirection make postprocessing of resolved expr_id's cumbersome
and error prone.

Avoid it.

In particular,
- make expr_id::resolved to refer directly to init_declarator,
  parameter_declaration_declarator and function definition entities
  rather than to function_definitions,
- drop direct_declarator_id::context and
- adapt the AST resolver implementation accordingly.

Signed-off-by: Nicolai Stange <nstange@suse.de>