ast: allow for '->' dereferences in an offset_member_designator
Although the GCC manual doesn't document it, it should be possible
to have '->' dereferences in a __builtin_offsetof()'s member designator:
as per C99 7.17(3), the offsetof macro should accept anything such that

  &(t.member-designator)

is an address constant if &t is.

If in the sequence of member designators, an array of some struct is
encountered, dereferencing it via '->' is perfectly valid:
consider
  struct A
  {
        struct { int b; } a[1];
  };
and
  offsetof(struct A, a->b)
for example.

Hence, allow '->' dereferences in a offset_member_designator.

Signed-off-by: Nicolai Stange <nstange@suse.de>