target_float: from_base10_exp(): cancel computation when the result is obvious
from_base10_exp() calculates the power of 10 to the given exponent.
However, for large values of this exponent, the computation is basically
unbounded although the resulting floating point constant  is known to
either be Zero or Inf. The GCC testsuite even has a testcase involving
floating point constants of the form 1.0e1000000.

Make from_base10_exp() check for these cases in each iteration of
the 10^exponent compiutation and return either Inf or Zero
as appropriate then.

Signed-off-by: Nicolai Stange <nstange@suse.de>