ast: designator_array: provide get_index_last() helper
In the usual case of a initializer's index designator not being
a range, the designator_array's ->_index_last is not set.

However, for inferring an initialized array's length, the maximum handled
index is of special interest. Implement the ->get_index_last() helper
to return either of ->_index_last or ->_index_first, depending on whether
the designator_array refers to an index range or to a single entry.

Signed-off-by: Nicolai Stange <nstange@suse.de>