Introduce constexpr_value for storing the values of constant expressions
C99 has the notion of constant expressions which can be evaluated at
compile time:
- integer constant expressions,
- arithmetic constant expression of either integer or float type,
- address constants.

In particular the first kind, integer constant expressions, can be used
as part of type definitions: for example for specifying bitfield widths
or array dimensions.

Apart from the type system, these values will also be needed when it
comes to implementing conditional inclusion.

Provide the constexpr_value class for storing the kind and the actual
value of expressions evaluating to constants.

Signed-off-by: Nicolai Stange <nstange@suse.de>