Introduce type system
Introduce a class hierarchy modeling the C99 type system.

All actual type classes inherit from the abstract common base class 'type'.
This class provides, based on the C++ double-dispatching idiom, a set of
virtual functions for determining if two given types are compatible in
the sense of C99. Also, type qualifiers, i.e. "const", "volatile" and
"restrict", if any, are recoreded in this superclass 'type'.

From the class 'type' root, two orthogonal inheritance tree branch off:
- One describing the possible declarator derivations of a type and
- the other one describing a type's arithmetic properties like integer
  signedness and width, if any.

Actual types can derive (indirectly) from both branches, i.e. have
a diamond class inheritance relationship.

For the declarator-derivations-describing branch, the topmost abstract
class is 'addressable_type'. For any type derived from this, a pointer
to that type forms a valid type. A counterexample would be bitfield types.
Classes derived from 'addressable_type' which are complete can occupy
some storage and thus, have associated alignment and size properties.

Derived from 'addressable_type' are 'returnable_type' and 'object_type'.
A 'returnable_type' can be returned from a function and an 'object_type'
can form an array's element type. For example, 'void' can be returned from
a function, but cannot form an array's element type. On the other hand,
an array type can again be an array element type of a "more outer" array,
but cannot returned from function.

The two inheritance paths are joined at the convenience class
'returnable_object_type' which derives from both, 'returnable_type' and
'object_type'. The majority of actual types like the standard integers,
floats, pointers and struct/union types are derived from this.

Function types are derived from 'function_type' which is an
'addressable_type'. There are two flavours: 'prototyped_function_type'
describing a function declared with a C99 prototype and
'unprototyped_function_type' resembling those functions declared without a
prototype, i.e. those whose parameter list is either empty or is an
identifier list lacking type information.

The inheritance branch describing arithmetic properties starts out with
the abstract base class 'artithmetic_type', itself being a child of
class 'type'.

Derived from that are the mutually exclusive abstract classes 'int_type'
and 'float_type'. Actual integer types derive from the former, the real
and complex float types from the latter each. Class 'int_type' provides
a virtual function for determining the result of the C99 "standard integer
promotion. Also, based on the C++ double-dispatching idiom, 'int_type'
provides virtual functions for evaluating the C99 "integer conversion"
result from two input integer types. Likewise, class 'arithmetic_type'
provides primitives for evaluating the C99 "arithmetic conversion".

Finally, it's noteworthy that there is an artifical type,
'builtin_func_type', not resembling any real C99 type. It's intent is
to act as a kind of dummy type representing the function designator part
in the invocations of builtin functions ("__builtin_...()").

For the type evaluation of builtin function invocation, it stores a
pointer to a reference to a factory producing instances of abstract type
'builtin_func'. Class 'builtin_func' provides an interface for actually
doing the type evaluations for a particular builtin function with the
parameters given in some invocation. While the class 'builtin_func'
interface gets introduced by this commit, actual implementations for
each of the known builtin functions as well as the binding to their
corresponding factories from the resolver code will be implemented by a
later one.

Objects of classes derived from class 'type' are always dynamically
allocated and managed by std::shared_ptr. The rationale is that
for example in
 int *a;
the type object representing the int can be shared by both, the
declaration_specifiers AST entity and the pointer type object for
its reference to the pointed-to type.

Most often, pointers to type objects will be pointers to one of the
abstract classes from the hierarchy described above. For example, an
expr AST entity can evaluate to a bitfield_type or to some
addressable_type and thus, its pointer to its resulting type object must
point to the most possible abstract base class, i.e. class 'type'.
When it comes to working with type objects, during type evaluation for
example, demultiplexing based on the objects' actual leaf class will
be neecessary. For example, the implementation for evaluating an
expression of the form
  pointer_type + int_type
will be different from the one for the evaluation of
  real_float_type + real_float_type.
In order to avoid explicit downcasts all over the place, the
handle_type<>() template is provided. It takes a callable and a number
of objects derived from class 'type', downcasts the objects to their
actual leaf class type and forwards them to the callable which can
then overload based on the actual types. This means that the callable is
expected to take exactly as many arguments as there are type arguments
passed to the handle_types<>() invocation. Type objects can be passed
either by reference or wrapped in std::shared_ptr<>s. In the first case,
the callable is expected to take a reference for the corresponding
argument, in the latter it will be passed std::shared_ptr<> to the
downcasted type. It should be noted that the already available
wrap_callables<> will be of great help for constructing an overloaded
callable from a number of C++ lambdas.

Finally, note that a number of type properties like storage size, native
alignment, signedness of plain char's and so depend on the ABI and the
compiler. In order to allow for dynamic specification of these properties,
define an abstract base class 'architecture'. It's current interface as
introduced by this commit by far exceeds what is required from the type
system itself, but will be needed when the AST type evaluation gets
implemented in a later commit.

Signed-off-by: Nicolai Stange <nstange@suse.de>