Define 'execution_charset_encoder' interface
With C99 and GCC extensions resp. C11, there are several kinds of string
or character literals:
 - "foo"
 - L"foo"
 - u8"foo"
 - u"foo"
 - U"foo"

Their correspoding character type depends on the compiler or ABI and
- the values of character literals (which qualify as C99 integer constant
  expressions) and
- the length of character arrays given by string literals
both depend on the execution charset which is also a property of the
compiler resp. ABI.

In conclusion, the evaluation of character and string literal depends
on the compiler resp. ABI and as such, should be made virtual through
the abstract 'architecture' class.

Define an abstract class 'execution_charset_encoder' providing facilities
to encode string literals as well as to supply character type information.
Introduce a pure virtual factory function to class architecture which is
in charge of creating such execution_charset_encoder instances based on
the string literal kind in question.

Signed-off-by: Nicolai Stange <nstange@suse.de>