ast, builtins: resolve expr_id to builtin_func::factory
For the evaluation of builtin function invocations, we want to have
the resolving code not only check whether a given identifier is builtin
or not, but to link the expr_id to that builtin's corresponding
builtin_func::factory.

Introduce support for linking to builtin_func::factory in
expr_id::resolved. Implement the static builtin_func::lookup() mapping
names to actual builtin_func::factory values. Note that this is a trivial
placeholder because there are no implementations of builtin_func yet.

Let the AST id resolver code link an expr_id to the result obtained from
builtin_func::lookup(), if any.

Signed-off-by: Nicolai Stange <nstange@suse.de>