ast: evaluate types and constant expressions
Implement type and constant expression evaluation on the AST. For this,
implement an intermediate base template class 'typed_ast_entity<>' where
AST entities associated with some type, such as declarators or expressions,
are to be derived from.

The type_ast_entity<> template is itself derived from the non-template
class _typed to make it easy to distinguish typed AST entities in a
AST walk. Moreover, class _typed declares a pure virtual function
evaluate_type() to be implemented for each typed AST entity.

Implement ast::evaluate() to conduct a DFS walk on the tree in the correct
order and invoke the type evaluations on each entity.
More specifically,
- (direct) declarators get processed in pre-order as their evaluation
  depends on the type of the "enclosing" (direct) declarator, if any.
- expressions get processed in post-order, because their evaluation
  depends on the types of the subexpressions, if any.

Finally, equip class expr with a pointer to a constexpr_value to receive
the result of constant expression evaluation, if any. Let the
::evaluate_type() for expression entities set it as appropriate.

Signed-off-by: Nicolai Stange <nstange@suse.de>