testsuite: make the 'cparse' testing utility run type evaluations
Now that the type evaluation has been implemented, let the 'cparse'
testing utility run it on all testcases.

All currently available testcases stem from the GCC testsuite and are
sometimes intentionally ill-formed because their intent is to test GCC's
diagnostic capabilities or certain features of its C dialect. The 'cparse'
utility would emit warnings for these cases and for proper testing, it
is desirable to verify that the 'cparse' output actually matches what
is expected. Make the DejaGnu testing driver, i.e.
testsuite/lib/cparse.exp, extract the expected 'cparse' output from
special comments in the testcases themselves and compare the actual output
against this. These special comments format is

  /* cp-out: <regexp> */

where <regexp> is to be applied to a single line of output. The n'th
such annotation gets applied to the n'th (non-blank) line of output and
each line of output is required to have a matching annotation or the
testcase is considered to have failed.

Add these "cp-out" annotations to all existing testcases where appropriate.

A couple of the currently available testcases fail because they make use of
features currently not supported by the type evaluation code like complex
floating literals or certain builtins. Introduce support to
testsuite/lib/cparse.exp for marking these as expected failures by
adding

  /* cp-xfail */

annotations to the testcases' source files.

Annotate the known failers with that tag. I'll decide later on whether or
not the missing feature is worth implementing or remove the testcase
otherwise.

Finally, note that there is one remaining "real" failer,
testsuite/cparse/gcc.c-torture/compile/20001018-1.c, which is due to
the fact that the type evaluation code currently does not complete types
by forming the composite type for redeclarations. This violates the C
standard and will have to get fixed. So I'm leaving this failer as-is.

Signed-off-by: Nicolai Stange <nstange@suse.de>