AST resolver: use tree structure for identifier linkage
Currently, the AST resolver simply lumps together all function_definition
and init_declarator's related by linkage in a circular list.

However, for type completion by composition, this structure is
insufficient, because declarations must be composited only with those
visible at their location.

For example, in

  void f(void)
  {
    extern int a[1];
  }

  int g(void)
  {
    extern int a[];
    sizeof(a); /* error */
  }

  int a[1];

the second declaration of a in g() would yield and incomplete type because
it must not get composited with either the first one in f() or the last one
at file scope.

Note that visibility is a transitive property here.

Link the declarations related by linkage and visibility such that a child
links to a parent if and only if the latter represents the innermost
declaration visible at the child.

This yields a forest of trees where
- the nodes within a tree are related by linkage and visibility and
- nodes from different trees are related by linkage only.

In the example from above, this forest would have three trees, each
consisting of a single root node only.

So far, relations between trees are not reflected in the data structures.
Introduce a new boolean member, '_target_is_visible', in the linkage
tree nodes. Set it to true for the inner tree links. Abuse the linkage
nodes' link members for linking together the root nodes from trees related
by linkage in a singly linked list and set '_target_is_visible' to false
for these.

Signed-off-by: Nicolai Stange <nstange@suse.de>