AST evaluation: implement type composition
C99 6.2.7 specifies that for redeclaration of types, the resulting type
is what it calls a "composite type".

Construct composite types for direct_declarator_id's related by linkage
to a previous declaration visible in their scope.

This is necessary to properly evaluate for example

  int a[1];
  void f(void)
  {
    extern int a[];
    sizeof(a);
  }

Annotate the testcases with the additional expected warnings about
incompatible redeclarations.

Signed-off-by: Nicolai Stange <nstange@suse.de>