preprocessor: do not normalize expanded macro arguments
The whitespace normalization of the expanded macro arguments in the
constructor of macro::instance is superfluous: the macro expansion code
itself does not rely on it and preprocessor::_create_macro_arg() and
and preprocessor::read_next_token() are now fully capable of doing all of
the whitespace normalization ever needed. Do not normalize the expanded
macro arguments.

Signed-off-by: Nicolai Stange <nstange@suse.de>