ast evaluator: evaluate preprocessor expressions as intmax_t
The C99 standard says that the conditional inclusion conditions must be
evaluated as intmax_t. Make the AST evaluator evaluate expression held in
an ast_pp_expr as intmax_t.

Introduce a boolean flag, _is_pp_expr, in the abstract base class ast.
Disable it from subclass ast_translation_unit and enable it from
ast_pp_expr.

Make the AST evaluation code query it for those expressions that are
eligible for preprocessor expressions, i.e. those where no casts and
identifiers are involved and which can possibly evaluate to integer
constant expressions. If set, evaluate these expressions as if their
resulting type had been (u)intmax_t. Note that this is not needed for
certain kinds of expressions made up from subexpressions, because standard
promotion and conversion rule apply: if the subexpressions are of type
(u)intmax_t already, then so will be the enclosing expression.

Signed-off-by: Nicolai Stange <nstange@suse.de>