AST: assign globally unique id's to struct or union types
In order to be able to refer to struct/union types in a globally unique
manner, assign them a unique id.

Derive it from the address of the first sou_decl_list_node in the list
of redeclarations.

Make the types code use that id for determining whether or not two struct
or union types are the same.

Drop sou_decl_list_node::get_declaration() which was being used for that
purpose only.

Signed-off-by: Nicolai Stange <nstange@suse.de>