AST: provide expr_member_deref::get_base() accessor
In order to allow the upcoming live patch creation code to build
dependencies for the pointer form of the member dereference operator,
i.e. to add a dependency to the pointed to struct or union type, provide
access to the base expression by providing the
expr_member_deref::get_base() accessor.

Signed-off-by: Nicolai Stange <nstange@suse.de>