AST: provide direct_declarator_func::get_parameter_declaration_list()
The upcoming live patch creation code needs to handle a function's
parameter types specially: for building their dependencies, they must
be treated as if never allocated.

For example, the mere existence of a

 struct A;
 void f(struct A);

declaration does not require a definition of struct A.

Provide access to a function declaration's parameters by means of the
direct_declarator_func::get_parameter_declaration_list() declarator.

Signed-off-by: Nicolai Stange <nstange@suse.de>