Support forward declarations of enums
C99 does not allow for forward declarations of the form

  enum A;

as is possible for structs and unions, c.f. 6.7.2.3(3):
  "A type specifier of the form
     enum identifier
   without an enumerator list shall only appear after the type it specifies
   is complete."

Also, the C99 standard is lacking analogues of 6.7.2.3(7) and (8) for
enums:
  "A declaration of the form
     struct-or-union identifier ;
   specifies a structure or union type and declares the identifier as a tag
   of that type."
and
  "If a type specifier of the form
     struct-or-union identifier
   occurs other than as part of one of the above forms, and no other
   declaration of the identifier as a tag is visible, then it declares an
   incomplete structure or union type, and declares the identifier as the
   tag of that type."

However, GCC extends these two latter rules to apply to enums as well, i.e.
it allows for forward declarations of enum tags. There are a couple of
sites in the Linux kernel sources taking advantage of this.

Support forward declarations of enum by applying the very same semantics
to enum tags as is the case for struct and unions.

To that end, replace the simple enum_ref -> enum_def link used so far by
a new enum_decl_link and enum_decl_list_node infrastructure in close
analogy as it is done for struct and unions with sou_decl_link and
sou_decl_list_node.

Make the AST resolver link the enum_decl_list_nodes and enum_decl_link
fields contained in enum_def resp. enum_ref instances.

Change the type system's enum_type class to refer to a enum_decl_list_node
rather than to the defining enum_def instance.

Finally, prepare the AST evalutator to deal with expressions of incomplete
integer or arithmetic type. This situation has been impossible before this
change and thus, it was safe to query those for features like e.g. the
integer width and signedness. Make the AST evaluator report an error in
all cases where an expression of incomplete enum_type is illicitly used.

Signed-off-by: Nicolai Stange <nstange@suse.de>