AST: linkage: enable tracking of last file scope declaration
Currently, all init_declarator and function_definition instances relating
to the same object are linked together in singly linked, non-circular list
maintained by the class linkage instances contained in these.

While going from a later (in file order) declaration to an earlier is
possible with this scheme, the reverse is not.

However, for declarations at file scope, C99 allows for the completion
of their types by later redeclarations, c.f. 6.2.7(4) and 6.9.2(2).

Thus, for identifiers declared at file scope and related by linkage, we'll
want to
- be able to track which of their resp. file scope declarations is the
  last one in file order and
- be able to get from earlier ones to later ones.

Extend class linkage to allow for the tracking of an identifier's last
file scope declaration.

More specifically, turn the non-circular list of linkage related
declarations into a circular one with the node representing the first
declaration pointing back to the last file scope declaration, if any.

To that end, introduce a new boolean member, "_is_first_in_chain", for
marking the class linkage instance possibly pointing back.

Make the linking primitives maintain it such that all but the first
class linkage instances have it set to false.

Modify the linking primitives to update the link back to the last file
scope declaration to point to the new link's source, if requested.

Finally, for file scope declarations which are not redeclarations, this
this backlink should point to themselves. Introduce the new
->set_first_at_file_scope() implementing this.

The intended usage is to call ->set_first_at_file_scope() on the first
file scope declaration seen and then to subsequently request an update of
the backlink each time a link from a file scope declaration to a prior one
is established.

Make the AST resolver do this.

Signed-off-by: Nicolai Stange <nstange@suse.de>