preprocessor: introduce class code_remark_raw
Currently, the preprocessor and the evaluation code make use of the same
class code_remark for reporting errors and warnings.

However, the needs with respect to specifying locations are different:
the preprocessor most often wants to give code locations in the form
of location within a certain file. The resolver and evaluation code on
the other hand operate on AST entities and thus, sequences of
preprocessing tokens, possibly spanning multiple input files.

Ultimately, the reporting facilities to be used from the code operating
on AST entities should be made capable of taking pp_tokens_range instances
for the location specification. In order to prepare for this, save the
current code_remark functionality for further usage by the preprocessor.
Introduce a new class code_remark_raw which at this point just mirrors
the current class code_remark.

Signed-off-by: Nicolai Stange <nstange@suse.de>