preprocessor: move implementation class macro::instance to preprocessor
The class implementing the macro expansion during preprocessing, class
macro::instance, is currently declared as a subclass of class macro, the
class used for representing macro definitions.

The motivation behind this was that the macro expansion implementation
relies on the sequence of replacement tokens held in class macro to be
normalized.

However, this made the communication between the other preprocessing code
and the macro expansion implementation cumbersome. One symptom is having
the _expansion_history member in class pp_token: it's only use is to avoid
macro recursions, i.e. it's really only used during expansion and gets
cleared before a pp_token gets passed to the higher levels.

Move the macro expansion code to the rest of the preprocessing code, namely
- move class macro::instance to the new class
  preprocessor::_macro_instance, private to class preprocessor
- and grant it access to the needed members of class macro by introdcing
  accessors as appropriate.

Signed-off-by: Nicolai Stange <nstange@suse.de>