preprocessor: rework tracking of dependencies on macro non-definedness
The current facility for tracking dependencies on macros' non-definedness,
used_macro_undefs, is based on macro_undef and thus, is only able to record
dependencies on earlier #undef directives.

However, as we are going to reshuffle the code a bit when writing the
final live patch, this is insufficient. Consider e.g.

  int a;
  #define a

With the current scheme, no dependency on any #undef would get recorded
for the the indentifier token 'a' in the declaration. However, if the
macro define happens to get moved or replicated in front of the declaration
in the resulting live patch, an extra #undef will have to get inserted
inbetween the #define and the declaration.

Switch to a name based scheme for handling the constraints on macro
non-definedness. Replace class used_macro_undef by the new class
macro_nondef_constraint.

Note that function-style macros will not get expanded if the referring
identifier is not followed by an opening parenthesis, e.g. in

 #define a()
 a b

the 'a' on the second line would not get expanded. Make class
macro_nondef_constraint store a string for naming the identifier which
must not have been defined as a macro. Moreover, make it contain a boolean,
'func_like_allowed' for indicating that function-style macro definitions of
the given identifier would still be Ok.

Replace the former container class used_macro_undefs by the new class
macro_nondef_constraints.

There is no need to store a macro_nondef_constraints instance with any
pp_token which is not the result from some macro expansion, because this
information can be derived from its contents.

However, for pp_token instances from macro instances, the situation is
more complicated. For example, there is a rule which prevents recursive
evaluation. In

  #define a a b
  a

the 'a' would be expanded once into the sequence 'a b' and expansion would
then stop. In particular it would be wrong to infer a macro non-definedness
constraint for the identifier 'a' from the presence of an identifier token
'a' in the preprocessing result. Now, consider an example involving the
concatenation operator:

  #define __concat(a, b) a ## b
  #define concat(a, b) __concat(a, b)
  concat(a, b)

As the identifiers 'a' and 'b' don't refer to any defined macro, the first
expansion would result in '__concat(a, b)' and after another expansion step
the final preprocessing output would read as 'ab'. Hence, neither of the
identifiers 'a' or 'b' can be found in the preprocessing output, yet
they must not refer to macro definitions in order to obtain equivalent
results.

Let the preprocessor record all encountered identifiers which did not
resolve to macro definitions during macro expansion at a
macro_nondef_constraints instance kept at the corresponding
macro_invocation.

Signed-off-by: Nicolai Stange <nstange@suse.de>