Implement a (macro) depreprocessor
Implement class depreprocessor for writing a sequence of possibly
transformed pp_tokens subsequences to a source_writer while retaining
macro invocations as possible.

Users are supposed to register a number of
depreprocessor::transformed_input_chunk instances representing such
pp_tokens subsequences and the transformation operations to be carried
out on each.

In addition to these transformed_input_chunks, any number of
pp_result::header_inclusion_child instances to be #included from the
written source may be added.

The key feature of class depreprocessor is to retain macro invocations
as possible and to emit #defines and #undefs if needed.

Currently, macro invocations can be kept if and only if their resulting
pp_tokens are to be copied over as a whole and unmodified to the output.

Signed-off-by: Nicolai Stange <nstange@suse.de>