preprocessor: allow for addition of root source files after construction
Currently, class preprocessor's constructor takes a header_inclusion_roots
instance containing the set of root source files to be processed and this
set can't be altered afterwards.

On the other hand, predefined macros (and undefs as well) are registered
with the preprocessor instance after its construction.

When it comes to parsing command lines, it will be beneficial to be able
to register everything with a class preprocessor instance by the time
it's seen. Introduce preprocessor::add_root_source() for registering
root source files after a preprocessor's construction. Remove the
header_inclusion_roots argument from its constructor.

Signed-off-by: Nicolai Stange <nstange@suse.de>